fix: gpu field lowest_earn_month type int to int64

+++ pkg/cep_ent/gpu.go
@@ -40,7 +40,7 @@ type Gpu struct {
 	// 内存
 	Memory int `json:"memory"`
 	// 保底最低月收益
-	LowestEarnMonth int `json:"lowest_earn_month"`
+	LowestEarnMonth int64 `json:"lowest_earn_month"`
 	// Edges holds the relations/edges for other nodes in the graph.
 	// The values are being populated by the GpuQuery when eager-loading is set.
 	Edges        GpuEdges `json:"edges"`
@@ -172,7 +172,7 @@ func (gp *Gpu) assignValues(columns []string, values []any) error {
 			if value, ok := values[i].(*sql.NullInt64); !ok {
 				return fmt.Errorf("unexpected type %T for field lowest_earn_month", values[i])
 			} else if value.Valid {
-				gp.LowestEarnMonth = int(value.Int64)
+				gp.LowestEarnMonth = value.Int64
 			}
 		default:
 			gp.selectValues.Set(columns[i], values[i])
